Add a command line flag to create an rr recording
Recently I've been asking people to send me rr recordings whenever
they encounter bugs, since it really streamlines the bug fixing process
(in particular, it cuts out the whole "get it to reproduce and capture
it in rr part of the process"). However, the process so far is not quite
trivial. You have to get rr from somewhere (in a version that isn't broken
for latest julia), you have to remember to pack it and then you have
to find a place to upload it. I put together a package
(https://github.com/Keno/BugReporting.jl) that takes care of all of these
steps, but the convenience isn't quite 100% yet. That's what this PR is
supposed to address. It adds `--bug=` command line flag that internally
loads up the BugReporting, but passes through all the command line
options, environment, etc. The idea is that to create a bug report
you just do all the things you usually do, but instead of launching
`julia`, you launch `julia --bug`. This should also work if you are
launching Julia from within a script, etc.

Demo: https://asciinema.org/a/JFdoGT2wlxBE7uyzM1ZjwgSA5